menu header style refactor
