+++ components/style.css
@@ -1,3 +1,12 @@
+:root{
+    --white: #e7eded;
+    --darkGreen: #16796f;
+    --lightGreen: #7cb7af;
+    --header: #9ca89e;
+}
+
+
+
 @media (max-width: 570px) {
    
   .extractedElement img{
@@ -296,7 +305,7 @@ h2{
     align-items: center;
     justify-content: space-between;
     position: fixed;
-    background: #9ca89e;
+    background: var(--header);
     width: 100%;
     top: 0;
     z-index: 99;
@@ -319,22 +328,44 @@ h2{
 
 .menu-overlay,
 .menu-overlay.hide {
+    background: #000;
     position: absolute;
     height: 100vh;
-    width: 245px;
+    width: 265px;
     left: 0;
     top: 0;
     visibility: hidden;
-    
-    background: white;
+    transform: translatex(-300px);
     z-index: 99999;
+    transition: 0.3s ease-in-out;
+    box-shadow: inset -4px 0 8px -6px rgba(0, 0, 0, 0.15);
+    display: flex;
+    flex-direction: column;
 }
 
 
 .menu-overlay:not(.hide) {
     visibility: visible;
+    transform: translatex(0px);
 }
 
+.menu-overlay:not(.hide) + svg{
+    color:white;                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
+}
+
+.menu-overlay .menu-links a {
+    padding: 10px;
+    color: var(--white);
+    font-size: 1.4rem;
+    font-weight: bold;
+}
+
+
+.menu-overlay .menu-links{
+    display: grid !important;
+    background:inherit;
+    border:none;
+}
 
 .menu-icon{
     display: none !important;
@@ -398,33 +429,33 @@ h2{
     transition: top 2s cubic-bezier(0.25, 1, 0.5, 1), opacity 0.4s ease;
 }
 .menu-links a{
-    color: #cadedf;
+    color: var(--white);
+    font-weight: bold;
 }
 .menu-links {
     display: flex;
     gap: 20px;
-    padding: 10px 26px;
     border-radius: 19px;
     box-shadow: 0 2px 6px rgba(0, 0, 2, 0.08);
     background: #7cb7af17;
-    border: #9ca89e 1px solid;
+    border: var(--header) 1px solid;
+    background: black;
+    padding: 12px 32px;
 }
+
 
 .menu-links a.active {
     cursor:default;
     position: relative;
-    color: #000;
+    
+    color: var(--darkGreen)
 }
-
-
-
-
 
 .menu-links a.active::after {
     content: "";
-    border-bottom: 1px solid #000;
+    border-bottom: 1px solid var(--darkGreen);
     position: absolute;
-    bottom: 0;
+    bottom: -2px;
     left: 0px;      
     width: 100%; 
     height: 1px;
@@ -454,12 +485,13 @@ h2{
     .menu-log-cta,
     .menu-links{
         display: none !important;
-        margin-top: 70px;
+        margin-top: 95px;
     }
 
-    .menu-overlay .menu-links{
-        display: grid !important;
-        
+    .menu-overlay .menu-log-cta{
+        display: block !important;    
+        width: fit-content;
+        margin: auto auto 30px auto;
     }
 
 }
